Seperating UserService by method into multiple Services
